Fix improper IndexOutOfBoundException
When updatingSnapshots, tests would still fail due to having more than expected
number of snapshots as we would throw an IndexOutOfBoundsException.
Instead we now return null. This is conceptually more correct as it will result
in a proper expected vs actual failure message, as well as fixing the updateSnapshots bug.